move login logic to base scraper
