+++ lib/scrapers/base-scraper.js
@@ -1,6 +1,12 @@
 import phantom from 'phantom';
 import { waitForUrls, NAVIGATION_ERRORS } from '../helpers/navigation';
-import { LOGIN_RESULT } from '../helpers/login';
+import { waitUntilElementFound, fillInput, clickButton } from '../helpers/elements-interactions';
+
+const LOGIN_RESULT = {
+  SUCCESS: 'success',
+  INVALID_PASSWORD: 'invalidPassword',
+  CHANGE_PASSWORD: 'changePassword',
+};
 
 class BaseScraper {
   async initialize() {
@@ -10,6 +16,18 @@ class BaseScraper {
 
   async exit() {
     await this.instance.exit();
+  }
+
+  async login(loginUrl, fields, submitButtonId, options, notifier) {
+    await this.page.open(loginUrl);
+    await waitUntilElementFound(this.page, submitButtonId);
+
+    await Promise.all(fields.map((field) => {
+      return fillInput(this.page, field.id, field.value);
+    }));
+
+    await clickButton(this.page, submitButtonId);
+    notifier.notify(options, 'logging in');
   }
 
   async analyzeLogin(possibleUrls) {
@@ -58,4 +76,4 @@ class BaseScraper {
   }
 }
 
-export default BaseScraper;
+export { BaseScraper, LOGIN_RESULT };
