+++ lib/scrapers/discount.js
@@ -1,8 +1,7 @@
 import moment from 'moment';
 
-import BaseScraper from './base-scraper';
+import { BaseScraper, LOGIN_RESULT } from './base-scraper';
 import ScraperNotifier from '../helpers/notifier';
-import { login as performLogin, LOGIN_RESULT } from '../helpers/login';
 import { waitForRedirect } from '../helpers/navigation';
 import fetch from '../helpers/fetch';
 import includeJQuery from '../helpers/imported-libs';
@@ -69,12 +68,22 @@ function getPossibleLoginUrls() {
   return urls;
 }
 
+function createLoginFields(credentials) {
+  return [
+    { id: 'tzId', value: credentials.id },
+    { id: 'tzPassword', value: credentials.password },
+    { id: 'aidnum', value: credentials.num },
+  ];
+}
+
 class DiscountScraper extends BaseScraper {
   async scrape(credentials, options = {}) {
     await super.initialize();
     notifier.notify(options, 'start scraping');
 
-    await this.login(credentials, options);
+    const loginUrl = `${BASE_URL}/LoginPages/Logon?multilang=he&t=P&pageKey=home&bank=d`;
+    await super.login(loginUrl, createLoginFields(credentials), 'submitButton', options, notifier);
+    await waitForRedirect(this.page);
 
     const loginResult = await super.analyzeLogin(getPossibleLoginUrls());
 
@@ -84,17 +93,6 @@ class DiscountScraper extends BaseScraper {
     await super.exit();
 
     return scrapeResult;
-  }
-
-  async login(credentials, options) {
-    const loginUrl = `${BASE_URL}/LoginPages/Logon?multilang=he&t=P&pageKey=home&bank=d`;
-    const loginFields = [
-      { id: 'tzId', value: credentials.id },
-      { id: 'tzPassword', value: credentials.password },
-      { id: 'aidnum', value: credentials.num },
-    ];
-    await performLogin(this.page, loginUrl, loginFields, 'submitButton', () => notifier.notify(options, 'logging in'));
-    await waitForRedirect(this.page);
   }
 }
