+++ lib/scrapers/leumi-card.js
@@ -1,9 +1,8 @@
 import buildUrl from 'build-url';
 import moment from 'moment';
 
-import BaseScraper from './base-scraper';
+import { BaseScraper, LOGIN_RESULT } from './base-scraper';
 import ScraperNotifier from '../helpers/notifier';
-import { login as performLogin, LOGIN_RESULT } from '../helpers/login';
 import { waitForRedirect, waitForPageLoad } from '../helpers/navigation';
 import { waitUntilElementFound } from '../helpers/elements-interactions';
 
@@ -141,13 +140,22 @@ function getPossibleLoginUrls() {
   return urls;
 }
 
+function createLoginFields(inputGroupName, credentials) {
+  return [
+    { id: `${inputGroupName}_txtUserName`, value: credentials.username },
+    { id: `${inputGroupName}_txtPassword`, value: credentials.password },
+  ];
+}
+
 class DiscountScraper extends BaseScraper {
   async scrape(credentials, options = {}) {
     await super.initialize();
 
     notifier.notify(options, 'start scraping');
 
-    await this.login(credentials, options);
+    const loginUrl = `${BASE_URL}/Anonymous/Login/CardHoldersLogin.aspx`;
+    const inputGroupName = 'PlaceHolderMain_CardHoldersLogin1';
+    await super.login(loginUrl, createLoginFields(inputGroupName, credentials), `${inputGroupName}_btnLogin`, options, notifier);
     await redirectOrDialog(this.page);
 
     const loginResult = await super.analyzeLogin(getPossibleLoginUrls());
@@ -158,16 +166,6 @@ class DiscountScraper extends BaseScraper {
     await super.exit();
 
     return scrapeResult;
-  }
-
-  async login(credentials, options) {
-    const loginUrl = `${BASE_URL}/Anonymous/Login/CardHoldersLogin.aspx`;
-    const inputGroupName = 'PlaceHolderMain_CardHoldersLogin1';
-    const loginFields = [
-      { id: `${inputGroupName}_txtUserName`, value: credentials.username },
-      { id: `${inputGroupName}_txtPassword`, value: credentials.password },
-    ];
-    await performLogin(this.page, loginUrl, loginFields, `${inputGroupName}_btnLogin`, () => notifier.notify(options, 'logging in'));
   }
 }
